+++ image processing/code/HarrisCorner.cpp
@@ -8,10 +8,10 @@
 namespace qlm
 {
 	template<ImageFormat frmt, pixel_t T>
-	std::vector<KeyPoint<int>> HarrisCorner(const Image<frmt, T>& in, const unsigned int block_size, const unsigned int filter_size, float k, float threshold, const  BorderMode<frmt, T>& border_mode)
+	std::vector<KeyPoint<int>> HarrisCorner(const Image<frmt, T>& in, const size_t block_size, const size_t filter_size, float k, float threshold, const  BorderMode<frmt, T>& border_mode)
 	{
-		const unsigned int width = in.Width();
-		const unsigned int height = in.Height();
+		const size_t width = in.Width();
+		const size_t height = in.Height();
 		// normalization factor
 		float scale = (1 << (filter_size - 1)) * block_size * std::numeric_limits<T>::max();
 		scale = 1.0 / scale;
@@ -203,6 +203,6 @@ namespace qlm
 	}
 
 	template std::vector<KeyPoint<int>>
-	HarrisCorner<ImageFormat::GRAY, uint8_t>(const Image<ImageFormat::GRAY, uint8_t>&, const unsigned int, const unsigned int, float, float, const BorderMode<ImageFormat::GRAY, uint8_t>&);
+	HarrisCorner<ImageFormat::GRAY, uint8_t>(const Image<ImageFormat::GRAY, uint8_t>&, const size_t, const size_t, float, float, const BorderMode<ImageFormat::GRAY, uint8_t>&);
 	
 }
