+++ image processing/code/common/image/image.cpp
@@ -5,7 +5,7 @@ namespace qlm
 {
 	// create image
 	template<ImageFormat frmt, pixel_t T>
-	void Image<frmt, T>::create(unsigned int img_width, unsigned int img_height, Pixel<frmt, T> pix)
+	void Image<frmt, T>::create(size_t img_width, size_t img_height, Pixel<frmt, T> pix)
 	{
 		width = img_width;
 		height = img_height;
@@ -25,7 +25,7 @@ namespace qlm
 	}
 
 	template<ImageFormat frmt, pixel_t T>
-	void Image<frmt, T>::create(unsigned int img_width, unsigned int img_height)
+	void Image<frmt, T>::create(size_t img_width, size_t img_height)
 	{
 		width = img_width;
 		height = img_height;
@@ -78,42 +78,42 @@ namespace qlm
 
 
 	// explicit instantiation
-	template void Image<ImageFormat::GRAY, uint8_t>::create(unsigned int, unsigned int, Pixel<ImageFormat::GRAY, uint8_t>);
-	template void Image<ImageFormat::GRAY, int16_t>::create(unsigned int, unsigned int, Pixel<ImageFormat::GRAY, int16_t>);
-	template void Image<ImageFormat::GRAY, float>::create(unsigned int, unsigned int, Pixel<ImageFormat::GRAY, float>);
+	template void Image<ImageFormat::GRAY, uint8_t>::create(size_t, size_t, Pixel<ImageFormat::GRAY, uint8_t>);
+	template void Image<ImageFormat::GRAY, int16_t>::create(size_t, size_t, Pixel<ImageFormat::GRAY, int16_t>);
+	template void Image<ImageFormat::GRAY, float>::create(size_t, size_t, Pixel<ImageFormat::GRAY, float>);
 
-	template void Image<ImageFormat::RGB, uint8_t>::create(unsigned int, unsigned int, Pixel<ImageFormat::RGB, uint8_t>);
-	template void Image<ImageFormat::RGB, int16_t>::create(unsigned int, unsigned int, Pixel<ImageFormat::RGB, int16_t>);
-	template void Image<ImageFormat::RGB, float>::create(unsigned int, unsigned int, Pixel<ImageFormat::RGB, float>);
+	template void Image<ImageFormat::RGB, uint8_t>::create(size_t, size_t, Pixel<ImageFormat::RGB, uint8_t>);
+	template void Image<ImageFormat::RGB, int16_t>::create(size_t, size_t, Pixel<ImageFormat::RGB, int16_t>);
+	template void Image<ImageFormat::RGB, float>::create(size_t, size_t, Pixel<ImageFormat::RGB, float>);
 
-	template void Image<ImageFormat::HLS, uint8_t>::create(unsigned int, unsigned int, Pixel<ImageFormat::HLS, uint8_t>);
-	template void Image<ImageFormat::HLS, int16_t>::create(unsigned int, unsigned int, Pixel<ImageFormat::HLS, int16_t>);
-	template void Image<ImageFormat::HLS, float>::create(unsigned int, unsigned int, Pixel<ImageFormat::HLS, float>);
+	template void Image<ImageFormat::HLS, uint8_t>::create(size_t, size_t, Pixel<ImageFormat::HLS, uint8_t>);
+	template void Image<ImageFormat::HLS, int16_t>::create(size_t, size_t, Pixel<ImageFormat::HLS, int16_t>);
+	template void Image<ImageFormat::HLS, float>::create(size_t, size_t, Pixel<ImageFormat::HLS, float>);
 
-	template void Image<ImageFormat::HSV, uint8_t>::create(unsigned int, unsigned int, Pixel<ImageFormat::HSV, uint8_t>);
-	template void Image<ImageFormat::HSV, int16_t>::create(unsigned int, unsigned int, Pixel<ImageFormat::HSV, int16_t>);
-	template void Image<ImageFormat::HSV, float>::create(unsigned int, unsigned int, Pixel<ImageFormat::HSV, float>);
+	template void Image<ImageFormat::HSV, uint8_t>::create(size_t, size_t, Pixel<ImageFormat::HSV, uint8_t>);
+	template void Image<ImageFormat::HSV, int16_t>::create(size_t, size_t, Pixel<ImageFormat::HSV, int16_t>);
+	template void Image<ImageFormat::HSV, float>::create(size_t, size_t, Pixel<ImageFormat::HSV, float>);
 
-	template void Image<ImageFormat::YCrCb, uint8_t>::create(unsigned int, unsigned int, Pixel<ImageFormat::YCrCb, uint8_t>);
-	template void Image<ImageFormat::YCrCb, int16_t>::create(unsigned int, unsigned int, Pixel<ImageFormat::YCrCb, int16_t>);
-	template void Image<ImageFormat::YCrCb, float>::create(unsigned int, unsigned int, Pixel<ImageFormat::YCrCb, float>);
+	template void Image<ImageFormat::YCrCb, uint8_t>::create(size_t, size_t, Pixel<ImageFormat::YCrCb, uint8_t>);
+	template void Image<ImageFormat::YCrCb, int16_t>::create(size_t, size_t, Pixel<ImageFormat::YCrCb, int16_t>);
+	template void Image<ImageFormat::YCrCb, float>::create(size_t, size_t, Pixel<ImageFormat::YCrCb, float>);
 
 	// -------------------------------------------------------------------------------------------------------------
-	template void Image<ImageFormat::GRAY, uint8_t>::create(unsigned int, unsigned int);
-	template void Image<ImageFormat::GRAY, int16_t>::create(unsigned int, unsigned int);
-	template void Image<ImageFormat::GRAY, float>::create(unsigned int, unsigned int);
-	template void Image<ImageFormat::RGB, uint8_t>::create(unsigned int, unsigned int);
-	template void Image<ImageFormat::RGB, int16_t>::create(unsigned int, unsigned int);
-	template void Image<ImageFormat::RGB, float>::create(unsigned int, unsigned int);
-	template void Image<ImageFormat::HLS, uint8_t>::create(unsigned int, unsigned int);
-	template void Image<ImageFormat::HLS, int16_t>::create(unsigned int, unsigned int);
-	template void Image<ImageFormat::HLS, float>::create(unsigned int, unsigned int);
-	template void Image<ImageFormat::HSV, uint8_t>::create(unsigned int, unsigned int);
-	template void Image<ImageFormat::HSV, int16_t>::create(unsigned int, unsigned int);
-	template void Image<ImageFormat::HSV, float>::create(unsigned int, unsigned int);
-	template void Image<ImageFormat::YCrCb, uint8_t>::create(unsigned int, unsigned int);
-	template void Image<ImageFormat::YCrCb, int16_t>::create(unsigned int, unsigned int);
-	template void Image<ImageFormat::YCrCb, float>::create(unsigned int, unsigned int);
+	template void Image<ImageFormat::GRAY, uint8_t>::create(size_t, size_t);
+	template void Image<ImageFormat::GRAY, int16_t>::create(size_t, size_t);
+	template void Image<ImageFormat::GRAY, float>::create(size_t, size_t);
+	template void Image<ImageFormat::RGB, uint8_t>::create(size_t, size_t);
+	template void Image<ImageFormat::RGB, int16_t>::create(size_t, size_t);
+	template void Image<ImageFormat::RGB, float>::create(size_t, size_t);
+	template void Image<ImageFormat::HLS, uint8_t>::create(size_t, size_t);
+	template void Image<ImageFormat::HLS, int16_t>::create(size_t, size_t);
+	template void Image<ImageFormat::HLS, float>::create(size_t, size_t);
+	template void Image<ImageFormat::HSV, uint8_t>::create(size_t, size_t);
+	template void Image<ImageFormat::HSV, int16_t>::create(size_t, size_t);
+	template void Image<ImageFormat::HSV, float>::create(size_t, size_t);
+	template void Image<ImageFormat::YCrCb, uint8_t>::create(size_t, size_t);
+	template void Image<ImageFormat::YCrCb, int16_t>::create(size_t, size_t);
+	template void Image<ImageFormat::YCrCb, float>::create(size_t, size_t);
 	// ------------------------------------------------------------------------------------------------------------
 	template Pixel<ImageFormat::GRAY, uint8_t> Image<ImageFormat::GRAY, uint8_t>::GetPixel(int, int, const BorderMode<ImageFormat::GRAY, uint8_t>&) const;
 	template Pixel<ImageFormat::GRAY, int16_t> Image<ImageFormat::GRAY, int16_t>::GetPixel(int, int, const BorderMode<ImageFormat::GRAY, int16_t>&) const;
